Delete empty posts when leaving editor route
Closes #4321
- Renamed editor-route-base to editor-base-route to fit the naming scheme of editor-base-view and editor-base-controller
- Moved duplicate willTransition functionality from editor/edit and editor/new to editor-base-route
- editor-base-route passes willTransition actions up the route chain
- EditorRoute deletes a post with no bodyScratch and a title of `(Untitled)` on willTransition
