+++ core/client/mixins/editor-base-route.js
@@ -23,6 +23,53 @@ var EditorRouteBase = Ember.Mixin.create(styleBody, ShortcutsRoute, loadingIndic
         // The actual functionality is implemented in utils/codemirror-shortcuts
         codeMirrorShortcut: function (options) {
             this.get('controller.codemirror').shortcut(options.type);
+        },
+
+        willTransition: function (transition) {
+            var controller = this.get('controller'),
+                isDirty = controller.get('isDirty'),
+
+                model = controller.get('model'),
+                isSaving = model.get('isSaving'),
+                isDeleted = model.get('isDeleted'),
+                modelIsDirty = model.get('isDirty');
+
+            this.send('closeSettingsMenu');
+
+            // when `isDeleted && isSaving`, model is in-flight, being saved
+            // to the server. when `isDeleted && !isSaving && !modelIsDirty`,
+            // the record has already been deleted and the deletion persisted.
+            //
+            // in either case  we can probably just transition now.
+            // in the former case the server will return the record, thereby updating it.
+            // @TODO: this will break if the model fails server-side validation.
+            if (!(isDeleted && isSaving) && !(isDeleted && !isSaving && !modelIsDirty) && isDirty) {
+                transition.abort();
+                this.send('openModal', 'leave-editor', [controller, transition]);
+                return;
+            }
+            // since the transition is now certain to complete..
+            window.onbeforeunload = null;
+
+            // remove model-related listeners created in editor-base-route
+            this.detachModelHooks(controller, model);
+            // Bubble the action for the EditorRoute to catch
+            return true;
+        },
+
+        // Delete empty posts locally, or autosaved posts on the server
+        // when leaving the editor route.
+        deleteEmptyPosts: function () {
+            var model = this.controller.get('model');
+            // Delete unsaved, empty records from the store
+            if (model.get('isNew')) {
+                model.deleteRecord();
+                return;
+            }
+            // A blank, autosaved post - kill it serverside as well.
+            if (model.get('title') === '(Untitled)' && !model.get('markdown')) {
+                return model.destroyRecord();
+            }
         }
     },
