+++ core/client/routes/editor.js
@@ -1,0 +1,13 @@
+var EditorRoute = Ember.Route.extend({
+    actions: {
+        willTransition: function (transition) {
+            // If leaving the editor routes,
+            // delete empty post model.
+            if (!/^editor/.test(transition.targetName)) {
+                this.send('deleteEmptyPosts');
+            }
+        }
+    }
+});
+
+export default EditorRoute;
