+++ core/client/routes/editor/edit.js
@@ -1,4 +1,4 @@
-import base from 'ghost/mixins/editor-route-base';
+import base from 'ghost/mixins/editor-base-route';
 import isNumber from 'ghost/utils/isNumber';
 import isFinite from 'ghost/utils/isFinite';
 
@@ -65,41 +65,8 @@ var EditorEditRoute = Ember.Route.extend(SimpleAuth.AuthenticatedRouteMixin, bas
         // used to check if anything has changed in the editor
         controller.set('previousTagNames', model.get('tags').mapBy('name'));
 
-        // attach model-related listeners created in editor-route-base
+        // attach model-related listeners created in editor-base-route
         this.attachModelHooks(controller, model);
-    },
-
-    actions: {
-        willTransition: function (transition) {
-            var controller = this.get('controller'),
-                isDirty = controller.get('isDirty'),
-
-                model = controller.get('model'),
-                isSaving = model.get('isSaving'),
-                isDeleted = model.get('isDeleted'),
-                modelIsDirty = model.get('isDirty');
-
-            this.send('closeSettingsMenu');
-
-            // when `isDeleted && isSaving`, model is in-flight, being saved
-            // to the server. when `isDeleted && !isSaving && !modelIsDirty`,
-            // the record has already been deleted and the deletion persisted.
-            //
-            // in either case  we can probably just transition now.
-            // in the former case the server will return the record, thereby updating it.
-            // @TODO: this will break if the model fails server-side validation.
-            if (!(isDeleted && isSaving) && !(isDeleted && !isSaving && !modelIsDirty) && isDirty) {
-                transition.abort();
-                this.send('openModal', 'leave-editor', [controller, transition]);
-                return;
-            }
-
-            // since the transition is now certain to complete..
-            window.onbeforeunload = null;
-
-            // remove model-related listeners created in editor-route-base
-            this.detachModelHooks(controller, model);
-        }
     }
 });
