+++ core/client/routes/editor/new.js
@@ -1,4 +1,4 @@
-import base from 'ghost/mixins/editor-route-base';
+import base from 'ghost/mixins/editor-base-route';
 
 var EditorNewRoute = Ember.Route.extend(SimpleAuth.AuthenticatedRouteMixin, base, {
     classNames: ['editor'],
@@ -27,46 +27,8 @@ var EditorNewRoute = Ember.Route.extend(SimpleAuth.AuthenticatedRouteMixin, base
         // used to check if anything has changed in the editor
         controller.set('previousTagNames', Ember.A());
 
-        // attach model-related listeners created in editor-route-base
+        // attach model-related listeners created in editor-base-route
         this.attachModelHooks(controller, model);
-    },
-
-    actions: {
-        willTransition: function (transition) {
-            var controller = this.get('controller'),
-                isDirty = controller.get('isDirty'),
-
-                model = controller.get('model'),
-                isNew = model.get('isNew'),
-                isSaving = model.get('isSaving'),
-                isDeleted = model.get('isDeleted'),
-                modelIsDirty = model.get('isDirty');
-
-            this.send('closeSettingsMenu');
-
-            // when `isDeleted && isSaving`, model is in-flight, being saved
-            // to the server. when `isDeleted && !isSaving && !modelIsDirty`,
-            // the record has already been deleted and the deletion persisted.
-            //
-            // in either case  we can probably just transition now.
-            // in the former case the server will return the record, thereby updating it.
-            // @TODO: this will break if the model fails server-side validation.
-            if (!(isDeleted && isSaving) && !(isDeleted && !isSaving && !modelIsDirty) && isDirty) {
-                transition.abort();
-                this.send('openModal', 'leave-editor', [controller, transition]);
-                return;
-            }
-
-            if (isNew) {
-                model.deleteRecord();
-            }
-
-            // since the transition is now certain to complete..
-            window.onbeforeunload = null;
-
-            // remove model-related listeners created in editor-route-base
-            this.detachModelHooks(controller, model);
-        }
     }
 });
